Proof of Concept - Scoping exercise II
This is PDF is the LaTex version of the PoC Scoping Exercise II required for FOAR705 week 4 assignment. It includes a narrowing down of a problem to solve for thesis (from scoping I exercise), a decomposition of that problem, identification of patterns and identification of possible solutions that can fix the problems identified and to make things easier for research to be conducted.