Learning_Journal_OpenRefine
Uploaded the zip and pdf files from Overleaf LaTex which includes all data carpentry exercises, including the latest OpenRefine. This document also includes a new separated error section (still in progress) and work undertaken on for Proof of Concept, and work carried out during the typesetting phase. This version will be uploaded for assessment for FOAR705 to show completion of work for OpenRefine component of the course.